fix(groupBadge): рефакторинг + поправил баг с добавлением волонтеров по одному
