+++ packages/admin/src/components/entities/group-badges/edit.tsx
@@ -3,7 +3,6 @@ import {
     Edit,
     EditButton,
     Form,
-    Modal,
     Space,
     Table,
     TextField,
@@ -11,30 +10,25 @@ import {
     useForm,
     useTable
 } from '@pankod/refine-antd';
-import { Input, Popconfirm, Row, Col, Divider } from 'antd';
-import { useList, useUpdateMany, IResourceComponentsProps } from '@pankod/refine-core';
+import { Col, Divider, Input, Popconfirm, Row } from 'antd';
+import { IResourceComponentsProps, useUpdateMany } from '@pankod/refine-core';
 import { DeleteOutlined } from '@ant-design/icons';
-import type { TableRowSelection } from 'antd/es/table/interface';
-import { useEffect, useMemo, useState } from 'react';
-import type { Key } from 'react';
+import { useEffect, useState } from 'react';
 
 import type { GroupBadgeEntity, VolEntity } from '~/interfaces';
-import { useMedia } from '~/shared/providers';
-import useVisibleDirections from '../vols/use-visible-directions';
 import { CreateEdit } from './common';
+import { AddVolunteerModal } from '~/components/entities/group-badges/add-volunteer-modal';
 
 const { Title } = Typography;
 
 export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
     const { mutate } = useUpdateMany();
-    const { isDesktop } = useMedia();
 
     const [volunteers, setVolunteers] = useState<Array<VolEntity & { markedDeleted: boolean; markedAdded: boolean }>>(
         []
     );
+
     const [openAdd, setOpenAdd] = useState(false);
-    const [searchText, setSearchText] = useState('');
-    const [selected, setSelected] = useState<Array<Key>>([]);
 
     const { formProps, id, saveButtonProps } = useForm<GroupBadgeEntity>({
         onMutationSuccess: () => {
@@ -59,11 +53,7 @@ export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
         }
     });
 
-    const { data: volunteersAll, isLoading: isVolunteersAllLoading } = useList<VolEntity>({
-        resource: 'volunteers'
-    });
-
-    const { setFilters, tableProps: currentVols } = useTable<VolEntity>({
+    const { setFilters, tableProps: currentVolsTableParams } = useTable<VolEntity>({
         resource: 'volunteers',
         initialFilter: [
             {
@@ -82,42 +72,25 @@ export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
         initialPageSize: 10000
     });
 
-    const visibleDirections = useVisibleDirections();
-
     useEffect(() => {
-        const prevVolunteersMap = new Map(volunteers.map((vol) => [vol.id, vol]));
-        const updatedVolunteers =
-            currentVols.dataSource?.map((vol) => ({
-                ...vol,
-                markedDeleted: prevVolunteersMap.get(vol.id)?.markedDeleted ?? false,
-                markedAdded: prevVolunteersMap.get(vol.id)?.markedAdded ?? false
-            })) ?? [];
-        setVolunteers(updatedVolunteers);
-    }, [currentVols.dataSource]);
+        console.log('called!');
 
-    const addVolunteers = (): void => {
-        //если волонтер уже был в списке, но помечен на удаление, убираем флаг удаления
-        const updatedVolunteers = volunteers.map((item) => ({
-            ...item,
-            markedDeleted: selected.includes(item.id) ? false : item.markedDeleted,
-            markedAdded: false
-        }));
+        // По идее, тут дополняем данные полями 'markedDeleted' и 'markedAdded' и записываем их в volunteers
+        setVolunteers((prevVolunteers) => {
+            const prevVolunteersMap = new Map(prevVolunteers.map((vol) => [vol.id, vol]));
 
-        // добавляем только тех волонтеров, которых не было в списке
-        const currentIds = volunteers.map((item) => item.id);
-        const newVolunteers =
-            volunteersAll?.data
-                .filter((item) => selected.includes(item.id) && !currentIds.includes(item.id))
-                .map((item) => ({ ...item, markedDeleted: false, markedAdded: true })) ?? [];
+            return (
+                currentVolsTableParams.dataSource?.map((vol) => ({
+                    ...vol,
+                    markedDeleted: prevVolunteersMap.get(vol.id)?.markedDeleted ?? false,
+                    markedAdded: prevVolunteersMap.get(vol.id)?.markedAdded ?? false
+                })) ?? []
+            );
+        });
+    }, [currentVolsTableParams.dataSource]);
 
-        setVolunteers([...updatedVolunteers, ...newVolunteers]);
-        setSelected([]);
-    };
-
-    const dataSource = useMemo(() => volunteers.filter((vol) => !vol.markedDeleted), [volunteers]);
-
-    const handleChangeInputValue = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const value = e.target.value;
+    const handleChangeInputValue = (event: React.ChangeEvent<HTMLInputElement>): void => {
+        const value = event.target.value;
         setFilters([
             {
                 field: 'group_badge',
@@ -130,43 +103,6 @@ export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
                 value
             }
         ]);
-    };
-
-    /**
-     * данные для попапа добавить волонтеров
-     */
-    const filteredData = useMemo(() => {
-        //todo если волонтер уже есть в volunteers но с флагом markedDeleted
-        const currentIds = volunteers.filter((item) => !item.markedDeleted).map((item) => item.id);
-        const filteredVols = volunteersAll?.data.filter(
-            (vol) => !visibleDirections || vol.directions?.some(({ id }) => visibleDirections.includes(id))
-        );
-
-        if (!searchText) {
-            return filteredVols?.filter((item) => !currentIds.includes(item.id));
-        }
-
-        const searchTextLower = searchText.toLowerCase();
-        return filteredVols?.filter((item) => {
-            const isSelected = selected.includes(item.id);
-            const isNotCurrent = !currentIds.includes(item.id);
-            const matchesSearch = [
-                item.name,
-                item.first_name,
-                item.last_name,
-                item.directions?.map(({ name }) => name).join(', ')
-            ]
-                .filter(Boolean)
-                .some((text) => text?.toLowerCase().includes(searchTextLower));
-
-            return isNotCurrent && (matchesSearch || isSelected);
-        });
-    }, [volunteersAll, volunteers, searchText, selected, visibleDirections]);
-
-    const rowSelection: TableRowSelection<VolEntity> = {
-        selectedRowKeys: selected,
-        onChange: setSelected,
-        type: 'checkbox'
     };
 
     const activeVolunteers = volunteers.filter((item) => !item.markedDeleted);
@@ -191,7 +127,7 @@ export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
                     </Space>
                 </Col>
             </Row>
-            <Table {...currentVols} dataSource={dataSource} rowKey='id'>
+            <Table {...currentVolsTableParams} dataSource={activeVolunteers} rowKey='id'>
                 <Table.Column dataIndex='name' key='name' title='Имя на бейдже' />
                 <Table.Column dataIndex='first_name' key='first_name' title='Имя' />
                 <Table.Column dataIndex='last_name' key='last_name' title='Фамилия' />
@@ -224,13 +160,13 @@ export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
                                 cancelText='Отмена'
                                 okType='danger'
                                 onConfirm={(): void => {
-                                    setVolunteers(
-                                        volunteers.find((item) => item.id === record.id)?.markedAdded
-                                            ? volunteers.filter((item) => item.id !== record.id)
-                                            : volunteers.map((vol) =>
+                                    setVolunteers((prevVolunteers) => {
+                                        return prevVolunteers.find((item) => item.id === record.id)?.markedAdded
+                                            ? prevVolunteers.filter((item) => item.id !== record.id)
+                                            : prevVolunteers.map((vol) =>
                                                   vol.id === record.id ? { ...vol, markedDeleted: true } : vol
-                                              )
-                                    );
+                                              );
+                                    });
                                 }}
                             >
                                 <Button icon={<DeleteOutlined />} danger size='small' />
@@ -239,34 +175,12 @@ export const GroupBadgeEdit: FC<IResourceComponentsProps> = () => {
                     )}
                 />
             </Table>
-            <Modal
-                title='Добавить волонтеров'
-                open={openAdd}
-                onOk={() => {
-                    addVolunteers();
-                    setOpenAdd(false);
-                }}
-                onCancel={() => setOpenAdd(false)}
-            >
-                <Input
-                    placeholder='Поиск...'
-                    value={searchText}
-                    onChange={(e) => setSearchText(e.target.value)}
-                    style={{ marginBottom: 16 }}
-                />
-                <Table
-                    rowSelection={rowSelection}
-                    dataSource={filteredData}
-                    rowKey='id'
-                    loading={isVolunteersAllLoading}
-                    size='small'
-                    pagination={{ pageSize: isDesktop ? 100 : 5, showSizeChanger: false, size: 'small' }}
-                >
-                    <Table.Column dataIndex='name' key='name' title='Имя на бейдже' ellipsis width='40%' />
-                    <Table.Column dataIndex='first_name' key='first_name' title='Имя' ellipsis />
-                    <Table.Column dataIndex='last_name' key='last_name' title='Фамилия' ellipsis />
-                </Table>
-            </Modal>
+            <AddVolunteerModal
+                setVolunteers={setVolunteers}
+                volunteers={volunteers}
+                setIsOpen={setOpenAdd}
+                isOpen={openAdd}
+            />
         </Edit>
     );
 };
